Revert "Revert "🔧 refactor gallery to remove eslint error (#68)" (#74)"
This reverts commit 158077c47539031d3b0c25e644c96674be09abd7.
